+++ ios/RNSentry.mm
@@ -9,7 +9,7 @@
 #import <Sentry/Sentry.h>
 #import <Sentry/PrivateSentrySDKOnly.h>
 #import <Sentry/SentryScreenFrames.h>
-#import "SentryOptions+RNOptions.h"
+#import <Sentry/SentryOptions+HybridSDKs.h>
 
 // Thanks to this guard, we won't import this header when we build for the old architecture.
 #ifdef RCT_NEW_ARCH_ENABLED
@@ -50,7 +50,35 @@ RCT_EXPORT_METHOD(initNativeSdk:(NSDictionary *_Nonnull)options
                   rejecter:(RCTPromiseRejectBlock)reject)
 {
     NSError *error = nil;
+    SentryOptions* sentryOptions = [self createOptionsWithDictionary:options error:&error];
+    if (error != nil) {
+        reject(@"SentryReactNative", error.localizedDescription, error);
+        return;
+    }
 
+    [SentrySDK startWithOptions:sentryOptions];
+
+#if TARGET_OS_IPHONE || TARGET_OS_MACCATALYST
+    BOOL appIsActive = [[UIApplication sharedApplication] applicationState] == UIApplicationStateActive;
+#else
+    BOOL appIsActive = [[NSApplication sharedApplication] isActive];
+#endif
+
+    // If the app is active/in foreground, and we have not sent the SentryHybridSdkDidBecomeActive notification, send it.
+    if (appIsActive && !sentHybridSdkDidBecomeActive && (PrivateSentrySDKOnly.options.enableAutoSessionTracking || PrivateSentrySDKOnly.options.enableWatchdogTerminationTracking)) {
+        [[NSNotificationCenter defaultCenter]
+            postNotificationName:@"SentryHybridSdkDidBecomeActive"
+            object:nil];
+
+        sentHybridSdkDidBecomeActive = true;
+    }
+
+    resolve(@YES);
+}
+
+- (SentryOptions *_Nullable)createOptionsWithDictionary:(NSDictionary *_Nonnull)options
+                                         error: (NSError *_Nonnull *_Nonnull) errorPointer
+{
     SentryBeforeSendEventCallback beforeSend = ^SentryEvent*(SentryEvent *event) {
         // We don't want to send an event after startup that came from a Unhandled JS Exception of react native
         // Because we sent it already before the app crashed.
@@ -74,10 +102,9 @@ RCT_EXPORT_METHOD(initNativeSdk:(NSDictionary *_Nonnull)options
     [mutableOptions removeObjectForKey:@"tracesSampleRate"];
     [mutableOptions removeObjectForKey:@"tracesSampler"];
 
-    SentryOptions *sentryOptions = [[SentryOptions alloc] initWithRNOptions:options didFailWithError:&error];
-    if (error) {
-        reject(@"SentryReactNative", error.localizedDescription, error);
-        return;
+    SentryOptions *sentryOptions = [[SentryOptions alloc] initWithDict:mutableOptions didFailWithError:errorPointer];
+    if (*errorPointer != nil) {
+        return nil;
     }
 
     if ([mutableOptions valueForKey:@"enableNativeCrashHandling"] != nil) {
@@ -91,36 +118,16 @@ RCT_EXPORT_METHOD(initNativeSdk:(NSDictionary *_Nonnull)options
     }
 
     // Enable the App start and Frames tracking measurements
-    if ([mutableOptions valueForKey:@"enableAutoPerformanceTracking"] != nil) {
-        BOOL enableAutoPerformanceTracking = (BOOL)[mutableOptions valueForKey:@"enableAutoPerformanceTracking"];
+    if ([mutableOptions valueForKey:@"enableAutoPerformanceTracing"] != nil) {
+        BOOL enableAutoPerformanceTracing = (BOOL)[mutableOptions valueForKey:@"enableAutoPerformanceTracing"];
 
-        PrivateSentrySDKOnly.appStartMeasurementHybridSDKMode = enableAutoPerformanceTracking;
+        PrivateSentrySDKOnly.appStartMeasurementHybridSDKMode = enableAutoPerformanceTracing;
 #if TARGET_OS_IPHONE || TARGET_OS_MACCATALYST
-        PrivateSentrySDKOnly.framesTrackingMeasurementHybridSDKMode = enableAutoPerformanceTracking;
+        PrivateSentrySDKOnly.framesTrackingMeasurementHybridSDKMode = enableAutoPerformanceTracing;
 #endif
     }
 
-    [SentrySDK startWithOptions:sentryOptions];
-
-#if TARGET_OS_IPHONE || TARGET_OS_MACCATALYST
-    BOOL appIsActive = [[UIApplication sharedApplication] applicationState] == UIApplicationStateActive;
-#else
-    BOOL appIsActive = [[NSApplication sharedApplication] isActive];
-#endif
-
-    // If the app is active/in foreground, and we have not sent the SentryHybridSdkDidBecomeActive notification, send it.
-    if (appIsActive && !sentHybridSdkDidBecomeActive && (PrivateSentrySDKOnly.options.enableAutoSessionTracking || PrivateSentrySDKOnly.options.enableOutOfMemoryTracking)) {
-        [[NSNotificationCenter defaultCenter]
-            postNotificationName:@"SentryHybridSdkDidBecomeActive"
-            object:nil];
-
-        sentHybridSdkDidBecomeActive = true;
-    }
-
-
-
-
-    resolve(@YES);
+    return sentryOptions;
 }
 
 - (void)setEventOriginTag:(SentryEvent *)event {
@@ -176,33 +183,27 @@ RCT_EXPORT_METHOD(fetchNativeDeviceContexts:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)
 {
     NSLog(@"Bridge call to: deviceContexts");
-    NSMutableDictionary<NSString *, id> *contexts = [NSMutableDictionary new];
+    __block NSMutableDictionary<NSString *, id> *contexts;
     // Temp work around until sorted out this API in sentry-cocoa.
     // TODO: If the callback isnt' executed the promise wouldn't be resolved.
     [SentrySDK configureScope:^(SentryScope * _Nonnull scope) {
-        NSDictionary<NSString *, id> *serializedScope = [scope serialize];
-        // Scope serializes as 'context' instead of 'contexts' as it does for the event.
-        NSDictionary<NSString *, id> *tempContexts = [serializedScope valueForKey:@"context"];
+        NSDictionary<NSString *, id>  *serializedScope = [scope serialize];
+        contexts = [serializedScope mutableCopy];
 
-        NSMutableDictionary<NSString *, id> *user = [NSMutableDictionary new];
-
-        NSDictionary<NSString *, id> *tempUser = [serializedScope valueForKey:@"user"];
-        if (tempUser != nil) {
-            [user addEntriesFromDictionary:[tempUser valueForKey:@"user"]];
-        } else {
-            [user setValue:PrivateSentrySDKOnly.installationID forKey:@"id"];
+        NSDictionary<NSString *, id>  *user = [contexts valueForKey:@"user"];
+        if (user == nil) {
+            [contexts
+                setValue:@{ @"id": PrivateSentrySDKOnly.installationID }
+                forKey:@"user"];
         }
-        [contexts setValue:user forKey:@"user"];
 
-        if (tempContexts != nil) {
-            [contexts setValue:tempContexts forKey:@"context"];
-        }
         if (PrivateSentrySDKOnly.options.debug) {
             NSData *data = [NSJSONSerialization dataWithJSONObject:contexts options:0 error:nil];
             NSString *debugContext = [[NSString alloc] initWithData:data encoding:NSUTF8StringEncoding];
             NSLog(@"Contexts: %@", debugContext);
         }
     }];
+
     resolve(contexts);
 }
 
@@ -388,12 +389,12 @@ RCT_EXPORT_METHOD(addBreadcrumb:(NSDictionary *)breadcrumb)
             sentryLevel = kSentryLevelFatal;
         } else if ([levelString isEqualToString:@"warning"]) {
             sentryLevel = kSentryLevelWarning;
-        } else if ([levelString isEqualToString:@"info"]) {
-            sentryLevel = kSentryLevelInfo;
+        } else if ([levelString isEqualToString:@"error"]) {
+            sentryLevel = kSentryLevelError;
         } else if ([levelString isEqualToString:@"debug"]) {
             sentryLevel = kSentryLevelDebug;
         } else {
-            sentryLevel = kSentryLevelError;
+            sentryLevel = kSentryLevelInfo;
         }
         [breadcrumbInstance setLevel:sentryLevel];
 
@@ -463,7 +464,7 @@ RCT_EXPORT_METHOD(enableNativeFramesTracking)
 {
     // Do nothing on iOS, this bridge method only has an effect on android.
     // If you're starting the Cocoa SDK manually,
-    // you can set the 'enableAutoPerformanceTracking: true' option and
+    // you can set the 'enableAutoPerformanceTracing: true' option and
     // the 'tracesSampleRate' or 'tracesSampler' option.
 }